requirements.txt
Django
Celery
django-celery-beat
django-celery-results
redis
python-decouple
undetected-chromedriver
gspread
google-auth
lxml[html_clean]
requests
gunicorn
uvicorn
cassandra-driver
beautifulsoup4
jupyter
selenium
fake-useragent
requests_html
slugify